+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 * :hammer: Color All method names that have started with make are changed to create
 * :hammer: Intent All method names that have started with make are changed to create
 * :fire: Remove Windowx isLandscape, isPortrait method
+* :fire: Displayx getDisplayMetrics rename to getScreenMetrics,
+    getDensity rename to getDensity,
+    getScreenDensityDpi rename to getDensityDpi,
+    isPortraitOrientation rename to isOrientationPortrait
+
 
 ### v0.2.3
 * :hammer: Network getIpAddress rename to getLocalIpAddress